rename 'commitsInLast90Days' to 'hasRecentCommits'" -s
